requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
cloudinary==1.32.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2.18
django-allauth==0.54.0
gunicorn==20.1.0
oauthlib==3.2.2
Pillow==9.5.0
psycopg2==2.9.6
PyJWT==2.6.0
pytesseract==0.3.10
python3-openid==3.2.0
pytz==2023.3
pyzbar==0.1.9
requests-oauthlib==1.3.1
sqlparse==0.4.3
urllib3==1.26.17
whitenoise==6.4.0
